--- src/components/List Group/index.ts
@@ -1,4 +1,0 @@
-import ListGroupWET from './ListGroup';
-import ListGroupItemWET from './ListGroupItem';
-
-export default { ListGroupWET, ListGroupItemWET };
