--- src/components/List Group/stories/List Group.intro.stories.mdx
@@ -1,23 +1,0 @@
-import { ArgsTable, Canvas, Meta } from '@storybook/addon-docs';
-
-import ListGroupWET from '../ListGroup';
-import ListGroupItemWET from 'src/components/List Group/ListGroupItem';
-
-<Meta title="Components/List Group/Intro" component={ListGroupWET} />
-
-# List Group
-
-List groups are a flexible and powerful component for displaying a series of content.
-Modify and extend them to support just about any content within.
-
-## Example
-
-<Canvas>
-  <ListGroupWET>
-    <ListGroupItemWET> Some stuff bro </ListGroupItemWET>
-  </ListGroupWET>
-</Canvas>
-
-## Available Props
-
-<ArgsTable />
