+++ src/components/ListGroup/index.ts
@@ -1,0 +1,12 @@
+import OriginListGroup from './ListGroup';
+import ListGroupItem from './ListGroupItem';
+
+export type ListGroupProps = typeof OriginListGroup & {
+  Item: typeof ListGroupItem;
+};
+
+const ListGroup = OriginListGroup as ListGroupProps;
+
+ListGroup.Item = ListGroupItem;
+
+export default ListGroup;
