+++ src/components/ListGroup/stories/ListGroup.intro.stories.mdx
@@ -1,0 +1,22 @@
+import { ArgsTable, Canvas, Meta } from '@storybook/addon-docs';
+
+import ListGroup from '../ListGroup';
+
+<Meta title="Components/List Group/Intro" component={ListGroup} />
+
+# List Group
+
+List groups are a flexible and powerful component for displaying a series of content.
+Modify and extend them to support just about any content within.
+
+### Example
+
+<Canvas>
+  <ListGroup>
+    <ListGroup.Item>11111</ListGroup.Item>
+    <ListGroup.Item>222222</ListGroup.Item>
+    <ListGroup.Item>33333</ListGroup.Item>
+  </ListGroup>
+</Canvas>
+
+### Available Props
