+++ src/components/index.ts
@@ -9,3 +9,4 @@ export { default as Label } from './Label';
 export { default as Panel } from './Panel';
 export { default as CodeBlock } from './CodeBlock';
 export { default as ProgressBar } from './ProgressBar';
+export { default as ListGroup } from './ListGroup';
